requirements.txt
langchain==0.3.3
langchain-community==0.3.1
langchain-chroma==0.1.4
langchain-crewai-tools==0.1.3
crewai==0.61.1
crewai-tools==0.9.3
chromadb==0.5.13
sentence-transformers==3.2.0
beautifulsoup4==4.12.3
pypdf==5.3.1
requests==2.32.3
gradio==5.6.1
python-dotenv==1.0.1
faiss-cpu==1.8.0.2
pandas==2.2.3
python-multipart==0.0.12

